requirements-core.txt
# CORE REQUIREMENTS - Essential packages only
# Tested and working with Python 3.9-3.13
# Install with: pip install -r requirements-core.txt

# MCP Protocol
mcp>=0.9.0

# Web Framework (for future web API)
fastapi>=0.104.0
uvicorn>=0.24.0

# Data Validation
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Database
sqlalchemy>=2.0.0
aiosqlite>=0.19.0

# HTTP Clients
httpx>=0.25.0
aiohttp>=3.9.0

# AI Integration
anthropic>=0.8.0

# Configuration & Environment
python-dotenv>=1.0.0
pyyaml>=6.0.1

# Logging
loguru>=0.7.2

# Async File Operations
aiofiles>=23.2.1

# Utilities
tenacity>=8.2.3
cachetools>=5.3.2

# Date/Time
python-dateutil>=2.8.2

# XML Processing (for Path of Building)
xmltodict>=0.13.0
defusedxml>=0.7.1

# Web Scraping (lightweight)
beautifulsoup4>=4.12.0
lxml>=4.9.3

# Performance
orjson>=3.9.10

# Windows Support
pywin32>=306; platform_system == "Windows"
